Fix compiler warning in chooseRandomValuesWithSum3

+++ my-own-problems/shared/random-utilities.h
@@ -120,7 +120,7 @@ inline std::vector<int64_t> chooseRandomValuesWithSum3(const int64_t numValues, 
                 - 1); // Since (indices[i] - indices[i - 1] >= 1), we must subtract 1 else all chosen numbers will always be >= minValue + 1!
     }
 
-    assert(chosenValues.size() == numValues);
+    assert(static_cast<int64_t>(chosenValues.size()) == numValues);
     assert(std::accumulate(chosenValues.begin(), chosenValues.end(), 0) == targetSum);
     assert(*min_element(chosenValues.begin(), chosenValues.end()) >= minValue);
